+++ tests/test_backends.py
@@ -38,7 +38,6 @@ from cogeo_mosaic.utils import get_footprints
 
 mosaic_gz = os.path.join(os.path.dirname(__file__), "fixtures", "mosaic.json.gz")
 mosaic_db = os.path.join(os.path.dirname(__file__), "fixtures", "mosaics.db")
-mosaic_bin = os.path.join(os.path.dirname(__file__), "fixtures", "mosaic.bin")
 mosaic_json = os.path.join(os.path.dirname(__file__), "fixtures", "mosaic.json")
 mosaic_jsonV1 = os.path.join(os.path.dirname(__file__), "fixtures", "mosaic_0.0.1.json")
 stac_page1 = os.path.join(os.path.dirname(__file__), "fixtures", "stac_p1.geojson")
@@ -91,14 +90,6 @@ def test_file_backend():
         assert len(mosaic.get_assets(150, 182, 9)) == 2
         assert len(mosaic.get_assets(147, 182, 12)) == 0
 
-    with MosaicBackend(mosaic_bin, backend_options={"gzip": True}) as mosaic:
-        assert isinstance(mosaic, FileBackend)
-        assert (
-            mosaic.mosaicid
-            == "24d43802c19ef67cc498c327b62514ecf70c2bbb1bbc243dda1ee075"
-        )
-        assert mosaic.quadkey_zoom == 7
-
     with MosaicBackend(mosaic_json) as mosaic:
         assert isinstance(mosaic, FileBackend)
         assert mosaic.quadkey_zoom == 7
@@ -135,12 +126,6 @@ def test_file_backend():
         with MosaicBackend("mosaic.json.gz", mosaic_def=mosaic_content) as mosaic:
             mosaic.write()
             with open("mosaic.json.gz", "rb") as f:
-                m = json.loads(_decompress_gz(f.read()))
-                assert m["quadkey_zoom"] == 7
-
-        with MosaicBackend("abinaryfile.bin", mosaic_def=mosaic_content) as mosaic:
-            mosaic.write(gzip=True)
-            with open("abinaryfile.bin", "rb") as f:
                 m = json.loads(_decompress_gz(f.read()))
                 assert m["quadkey_zoom"] == 7
 
@@ -291,31 +276,19 @@ def test_s3_backend(session):
     session.reset_mock()
 
     session.return_value.client.return_value.head_object.return_value = False
-    with MosaicBackend("s3://mybucket/00000", mosaic_def=mosaic_content) as mosaic:
-        assert isinstance(mosaic, S3Backend)
-        mosaic.write(gzip=True)
-    session.return_value.client.return_value.get_object.assert_not_called()
-    session.return_value.client.return_value.head_object.assert_called_once()
-    kwargs = session.return_value.client.return_value.put_object.call_args[1]
-    assert kwargs["Bucket"] == "mybucket"
-    assert kwargs["Key"] == "00000"
-    assert MosaicJSON(**json.loads(_decompress_gz(kwargs["Body"])))
-    session.reset_mock()
-
-    session.return_value.client.return_value.head_object.return_value = False
-    with MosaicBackend("s3://mybucket/00000", mosaic_def=mosaic_content) as mosaic:
+    with MosaicBackend("s3://mybucket/00000.json", mosaic_def=mosaic_content) as mosaic:
         assert isinstance(mosaic, S3Backend)
         mosaic.write()
     session.return_value.client.return_value.get_object.assert_not_called()
     session.return_value.client.return_value.head_object.assert_called_once()
     kwargs = session.return_value.client.return_value.put_object.call_args[1]
     assert kwargs["Bucket"] == "mybucket"
-    assert kwargs["Key"] == "00000"
+    assert kwargs["Key"] == "00000.json"
     assert MosaicJSON(**json.loads(kwargs["Body"]))
     session.reset_mock()
 
     session.return_value.client.return_value.head_object.return_value = True
-    with MosaicBackend("s3://mybucket/00000", mosaic_def=mosaic_content) as mosaic:
+    with MosaicBackend("s3://mybucket/00000.json", mosaic_def=mosaic_content) as mosaic:
         assert isinstance(mosaic, S3Backend)
         with pytest.raises(MosaicExistsError):
             mosaic.write()
@@ -325,7 +298,7 @@ def test_s3_backend(session):
     session.reset_mock()
 
     session.return_value.client.return_value.head_object.return_value = True
-    with MosaicBackend("s3://mybucket/00000", mosaic_def=mosaic_content) as mosaic:
+    with MosaicBackend("s3://mybucket/00000.json", mosaic_def=mosaic_content) as mosaic:
         assert isinstance(mosaic, S3Backend)
         mosaic.write(overwrite=True)
     session.return_value.client.return_value.get_object.assert_not_called()
@@ -500,7 +473,7 @@ def test_stac_backend(post):
         ]
 
     with STACBackend(
-        "https://a_stac.api/search", {}, 8, 14, backend_options={"max_items": 8}
+        "https://a_stac.api/search", {}, 8, 14, stac_api_options={"max_items": 8}
     ) as mosaic:
         assert mosaic._backend_name == "STAC"
         assert isinstance(mosaic, STACBackend)
@@ -532,7 +505,7 @@ def test_stac_backend(post):
         ]
 
     with STACBackend(
-        "https://a_stac.api/search", {}, 8, 14, backend_options={"max_items": 15}
+        "https://a_stac.api/search", {}, 8, 14, stac_api_options={"max_items": 15}
     ) as mosaic:
         assert mosaic._backend_name == "STAC"
         assert isinstance(mosaic, STACBackend)
@@ -556,7 +529,7 @@ def test_stac_backend(post):
         ]
 
     with STACBackend(
-        "https://a_stac.api/search", {}, 8, 14, backend_options={"max_items": 15}
+        "https://a_stac.api/search", {}, 8, 14, stac_api_options={"max_items": 15}
     ) as mosaic:
         with pytest.raises(NotImplementedError):
             mosaic.write()